Unit Of Work Pattern

Maintains a list of objects affected by a business transaction and coordinates the writing out of changes and the resolution of concurrency problems. Fowler (2002) , p. 184.

References

Fowler M (2002). Patterns of enterprise application architecture. Addison-Wesley Longman Publishing Co., Inc. ISBN 0321127420.

See also

Super class

R6P::Singleton -> Repository

Methods

Public methods

AbstractUnitOfWork$new()

AbstractUnitOfWork$finalize()

AbstractUnitOfWork$enter()

AbstractUnitOfWork$exit()

AbstractUnitOfWork$commit()

AbstractUnitOfWork$rollback()

Inherited methods

Method new()

Instantiate an object

Usage

AbstractUnitOfWork$new()

Method finalize()

Remove an object

Usage

AbstractUnitOfWork$finalize()

Method enter()

Start keeping track of changes that affect a database.

Usage

AbstractUnitOfWork$enter()

Method exit()

Figure out everything that needs to be done to alter the database as a result of your work.

Usage

AbstractUnitOfWork$exit()

Method commit()

Register the object with the Unit of Work for changes.

Usage

AbstractUnitOfWork$commit()

Method rollback()

Clear changes.

Usage

AbstractUnitOfWork$rollback()