Singleton Pattern

Ensure a class only has one instance, and provide a global point of access to it.

Details

Intent

Caution: Singletons can be a problem in multi-threaded applications, especially when they manipulate mutable data.

Tip: Singletons work well for immutable data, such as reading from some data source, since anything that can’t change isn’t going to run into thread clash problems.

Motivation

Examples

Counter

Counter <<- R6::R6Class(inherit = Singleton, public = list(
    count = 0,
    add_1 = function(){self$count = self$count + 1; invisible(self)}
))

counter_A <- Counter$new()
counter_A$count
#> [1] 0
counter_A$add_1()$count
#> [1] 1

counter_B <- Counter$new()
counter_B$count
#> [1] 1

identical(counter_A, counter_B)
#> [1] TRUE

Data Transfer Object (DTO)

DTO <- R6::R6Class(classname = "DTO", inherit = Singleton, public = list(
    con = NULL,
    initialize = function(){
        super$initialize()
        self$establish_connection()
    },
    establish_connection = function(){
        self$con <- DBI::dbConnect(RSQLite::SQLite(), dbname = ":memory:")
        return(self)
    },
    dbSendQuery = function(statement){
        res <- DBI::dbSendQuery(self$con, statement)
        return(res)
    }
))

database_A <- DTO$new()
database_B <- DTO$new()
identical(database_A, database_B)

References

References

https://en.wikipedia.org/wiki/Singleton_pattern

Methods

Public methods

Singleton$new()

Method new()

Create or retrieve an object

Usage

Singleton$new()