Repository Pattern

Mediates between the domain and data mapping layers using a collection-like interface for accessing domain objects.

Details

With Repository, in-memory objects do not need to know whether there is a database present or absent, they need no SQL interface code, and certainly no knowledge of the database schema.

How It Works

Repository isolates domain objects from details of the database access code

Repository concentrates code of query construction

Repository helps to minimize duplicate query logic

In R, the simplest form of Repository encapsulates data.frame entries persisted in a data store and the operations performed over them, providing a more object-oriented view of the persistence layer. From the caller point of view, the location (locally or remotely), the technology and the interface of the data store are obscured.

When to Use It

In situations with multiple data sources.

In situations where the real data store, the one that is used in production, is remote. This allows you to implement a Repository mock with identical queries that runs locally. Then, the mock could be used during development and testing. The mock itself may comprise a sample of the real data store or just fake data.

In situations where the real data store doesn’t exist. Implementing a mock Repository allows you to defer immature decisions about the database technology and/or defer its deployment. In this way, the temporary solution allows you to focus the development effort on the core functionality of the application.

In situations where using SQL queries can be represented by meaningful names. For example Repository$get_efficient_cars() = SELECT * FROM mtcars WHERE mpg > 20

When building stateless microservices.

Implementations

The code of the abstract base class of Repository is

AbstractRepository <- R6::R6Class("Repository", inherit = Singleton, cloneable = FALSE, public = list(
    #' @description Instantiate an object
    initialize = function() exceptions$not_implemented_error(),
    #' @description Add an element to the Repository.
    add = function(key, value) exceptions$not_implemented_error(),
    #' @description Delete an element from the Repository.
    del = function(key) exceptions$not_implemented_error(),
    #' @description Retrieve an element from the Repository.
    get = function(key) exceptions$not_implemented_error()
))

The minimal implementing of AbstractRepository requires you to define four functions:

initialize establishes a database connection of some sort;

add adds one or more domain objects into the database;

del deletes one or more domain objects from the database; and

get retrieve one or more domain objects from the database.

Importantly, any Repository implementation happens in context. The following implementations consider a repostory of car models with their specifications.

From the caller perspective, both implementation behave identically – they have the same queries. Under the hood, the two implementations employ different storage approaches.

Example: In-Memory Storage Implementation with collections

Tip: In-Memory implementations are a temporal solution that is good for testing and rapid prototyping.

In-Memory implementations contribute to rapid development because:

They can be used before you establish/get access to a real database.

They are fast to establish in comparison to DBMS

In-Memory implementations are useful during testing because they are independent of the real database (if any), that means:

They start as empty storage allowing the programmer to test specific behaviour of the caller.

They don’t make unintended changes in the real database

Caution: In-Memory implementations are not recommended during the production stage. In-Memory storage is lost when a session is rebooted. You should think about what are the ramifications of losing all the data put into storage.

See more information about collections::dict in appendix.

InMemoryRepository <- R6::R6Class(
  classname = "Repository", inherit = AbstractRepository, public = list(
    initialize = function() {private$cars <- collections::dict()},
    add = function(key, value){private$cars$set(key, value); invisible(self)},
    del = function(key){private$cars$remove(key); invisible(self)},
    get = function(key){return(private$cars$get(key, default = private$NULL_car))}
  ), private = list(
    NULL_car = cbind(uid  = NA_character_, datasets::mtcars)[0,],
    cars = NULL
  ))

Adding customised operations

InMemoryRepository$set("public", "get_all_cars", overwrite = TRUE, function(){
  private$cars$values() %>% dplyr::bind_rows()
})

mtcars <- datasets::mtcars %>% tibble::rownames_to_column("uid")
head(mtcars, 2)
#>             uid mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> 1     Mazda RX4  21   6  160 110  3.9 2.620 16.46  0  1    4    4
#> 2 Mazda RX4 Wag  21   6  160 110  3.9 2.875 17.02  0  1    4    4

repository <- InMemoryRepository$new()

repository$add(key = mtcars[1, "uid"], value = mtcars[1,])
repository$add(key = mtcars[2, "uid"], value = mtcars[2,])

repository$get(key = "Mazda RX4")
#>         uid mpg cyl disp  hp drat   wt  qsec vs am gear carb
#> 1 Mazda RX4  21   6  160 110  3.9 2.62 16.46  0  1    4    4

repository$get_all_cars()
#>             uid mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> 1     Mazda RX4  21   6  160 110  3.9 2.620 16.46  0  1    4    4
#> 2 Mazda RX4 Wag  21   6  160 110  3.9 2.875 17.02  0  1    4    4

repository$del(key = "Mazda RX4")
repository$get(key = "Mazda RX4")
#>  [1] uid  mpg  cyl  disp hp   drat wt   qsec vs   am   gear carb
#> <0 rows> (or 0-length row.names)

Talk about each line

Talk about additional functions

Example: Persistent Storage Implementation with DBI

PersistentRepository <- R6::R6Class(
  classname = "Repository", inherit = AbstractRepository, public = list(
    #' @param immediate (`logical`) Should queries be committed immediately?
    initialize = function(immediate = TRUE){
      private$immediate <- immediate
      private$conn <- DBI::dbConnect(RSQLite::SQLite(), dbname = ":memory:")
      DBI::dbCreateTable(private$conn, "mtcars", private$NULL_car)
    },
    add = function(key, value){
      car <- private$NULL_car %>% tibble::add_row(value)
      self$del(key = key)
      DBI::dbAppendTable(private$conn, "mtcars", car)
      invisible(self)
    },
    del = function(key){
      statement <- paste0("DELETE FROM mtcars WHERE uid = '", key, "'")
      DBI::dbExecute(private$conn, statement, immediate = private$immediate)
      invisible(self)
    },
    get = function(key){
      statement <- paste0("SELECT * FROM mtcars WHERE uid = '", key, "'")
      result <- DBI::dbGetQuery(private$conn, statement)
      if(nrow(result) == 0) return(private$NULL_car) else return(result)
    }
  ), private = list(
    NULL_car = cbind(uid  = NA_character_, datasets::mtcars)[0,],
    immediate = NULL,
    conn = NULL)
)

Adding customised operations

PersistentRepository$set("public", "get_all_cars", overwrite = TRUE, function(){
  statement <- "SELECT * FROM mtcars"
  result <- DBI::dbGetQuery(private$conn, statement)
  if(nrow(result) == 0) return(NULL) else return(result)
})

mtcars <- datasets::mtcars %>% tibble::rownames_to_column("uid")
head(mtcars, 2)
#>             uid mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> 1     Mazda RX4  21   6  160 110  3.9 2.620 16.46  0  1    4    4
#> 2 Mazda RX4 Wag  21   6  160 110  3.9 2.875 17.02  0  1    4    4

repository <- PersistentRepository$new()

repository$add(key = mtcars[1, "uid"], value = mtcars[1,])
repository$add(key = mtcars[2, "uid"], value = mtcars[2,])

repository$get(key = "Mazda RX4")
#>         uid mpg cyl disp  hp drat   wt  qsec vs am gear carb
#> 1 Mazda RX4  21   6  160 110  3.9 2.62 16.46  0  1    4    4

repository$get_all_cars()
#>             uid mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> 1     Mazda RX4  21   6  160 110  3.9 2.620 16.46  0  1    4    4
#> 2 Mazda RX4 Wag  21   6  160 110  3.9 2.875 17.02  0  1    4    4

repository$del(key = "Mazda RX4")
repository$get(key = "Mazda RX4")
#>  [1] uid  mpg  cyl  disp hp   drat wt   qsec vs   am   gear carb
#> <0 rows> (or 0-length row.names)

Further Reading

Repository at Martin Fowler Blog

Super class

R6P::Singleton -> Repository

Methods

Public methods

AbstractRepository$new()

AbstractRepository$add()

AbstractRepository$del()

AbstractRepository$get()

Inherited methods

Method new()

Instantiate an object

Usage

AbstractRepository$new()

Method add()

Add an element to the Repository.

Usage

AbstractRepository$add(key, value)

Arguments

key

(character) Name of the element.

value

(?) Value of the element. Note: The values in the Repository are not necessarily of the same type. That depends on the implementation of AbstractRepository.

Method del()

Delete an element from the Repository.

Usage

AbstractRepository$del(key)

Arguments

key

(character) Name of the element.

Method get()

Retrieve an element from the Repository.

Usage

AbstractRepository$get(key)

Arguments

key

(character) Name of the element.