Repository Pattern

Mediates between the domain and data mapping layers using a collection-like interface for accessing domain objects.

Details

How It Works

See [@Fowler2002, p. 322]

NULL

AbstractRepository <- R6::R6Class("Repository", inherit = Singleton, cloneable = FALSE, public = list(
    #' @description Instantiate an object
    initialize = function() exceptions$not_implemented_error(),
    #' @description Add an element to the Repository.
    add = function(key, value) exceptions$not_implemented_error(),
    #' @description Delete an element from the Repository.
    del = function(key) exceptions$not_implemented_error(),
    #' @description Retrieve an element from the Repository.
    get = function(key) exceptions$not_implemented_error()
))

Implementations

The minimal implementing of AbstractRepository requires you to define four functions:

initialize,

add,

del, and

get.

Importantly, any Repository implementation happens in context. The following implementations consider a repostory of car models with their specifications.

In-Memory Storage with collections

Tip: In-Memory implementations are a temporal solution that is good for testing and rapid prototyping.

In-Memory implementations contribute to rapid development because:

They can be used before you establish/get access to a real database.

They are fast to establish in comparison to DBMS

In-Memory implementations are useful during testing because they are independent of the real database (if any), that means:

They start as empty storage allowing the programmer to test specific behaviour of the caller.

They don’t make unintended changes in the real database

Caution: In-Memory implementations are not recommended during the production stage. In-Memory storage are lost when a session is rebooted. You should think about what are the ramifications of losing all the data put into storage.

See more information about collections::dict in appendix.

InMemoryRepository <- R6::R6Class(
  classname = "Repository", inherit = AbstractRepository, public = list(
    initialize = function() private$cars <- collections::dict(),
    add = function(key, value){private$cars$set(key, value); invisible(self)},
    del = function(key){private$cars$remove(key); invisible(self)},
    get = function(key){return(private$cars$get(key, default = NULL))}
  ), private = list(
    cars = NULL)
)

Adding customised operations

InMemoryRepository$set("public", "get_all_cars", overwrite = TRUE, function(){
  private$cars$values() %>% dplyr::bind_rows()
})

repository <- InMemoryRepository$new()

repository$add(key = rownames(mtcars[1,]), value = mtcars[1,])
repository$add(key = rownames(mtcars[2,]), value = mtcars[2,])

repository$get(key = rownames(mtcars[1,]))
#>           mpg cyl disp  hp drat   wt  qsec vs am gear carb
#> Mazda RX4  21   6  160 110  3.9 2.62 16.46  0  1    4    4

repository$get_all_cars()
#>               mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4      21   6  160 110  3.9 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag  21   6  160 110  3.9 2.875 17.02  0  1    4    4

repository$del(key = rownames(mtcars[1,]))
repository$get(key = rownames(mtcars[1,]))
#> NULL

Talk about each line

Talk about additional functions

Persistent Storage with DBI

PersistentRepository <- R6::R6Class(
    classname = "Repository", inherit = AbstractRepository, public = list(
        #' @param immediate (`logical`) Should queries be committed immediately?
        initialize = function(immediate = TRUE){
            private$immediate <- immediate
            private$conn <- DBI::dbConnect(RSQLite::SQLite(), dbname = ":memory:")
            private$NULL_car <-
                datasets::mtcars[0,] %>%
                tibble::add_column(uid = NA_character_, .before = 0)
            DBI::dbCreateTable(private$conn, "mtcars", private$NULL_car)
        },
        add = function(key, value){
            value <- private$NULL_car %>% tibble::add_row(uid = key, value)
            self$del(key = key)
            DBI::dbAppendTable(private$conn, "mtcars", value)
            invisible(self)
        },
        del = function(key){
            statement <- paste0("DELETE FROM mtcars WHERE uid = '", key, "'")
            DBI::dbExecute(private$conn, statement, immediate = private$immediate)
            invisible(self)
        },
        get = function(key){
            statement <- paste0("SELECT * FROM mtcars WHERE uid = '", key, "'")
            result <- DBI::dbGetQuery(private$conn, statement)
            if(nrow(result) == 0) return(NULL) else return(result)
        }
    ), private = list(
        immediate = NULL,
        NULL_car = NULL,
        conn = NULL)
)

Adding customised operations

PersistentRepository$set("public", "get_all_cars", overwrite = TRUE, function(){
    statement <- "SELECT * FROM mtcars"
    result <- DBI::dbGetQuery(private$conn, statement)
    if(nrow(result) == 0) return(NULL) else return(result)
})

repository <- PersistentRepository$new()

repository$add(key = rownames(mtcars[1,]), value = mtcars[1,])
repository$add(key = rownames(mtcars[2,]), value = mtcars[2,])

repository$get(key = rownames(mtcars[1,]))
#>         uid mpg cyl disp  hp drat   wt  qsec vs am gear carb
#> 1 Mazda RX4  21   6  160 110  3.9 2.62 16.46  0  1    4    4

repository$get_all_cars()
#>             uid mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> 1     Mazda RX4  21   6  160 110  3.9 2.620 16.46  0  1    4    4
#> 2 Mazda RX4 Wag  21   6  160 110  3.9 2.875 17.02  0  1    4    4

repository$del(key = rownames(mtcars[1,]))
repository$get(key = rownames(mtcars[1,]))
#> NULL

References

[@Fowler2002, p. 322]: R:@Fowler2002,%20p.%20322 [!snippet %>% stringr::str_detect(“^(#’|#)”)]: R:!snippet%20%25%3E%25%20stringr::str_detect(%22%5E(#’%7C#)%20%22) [collections]: R:%60collections%60 [appendix]: R:appendix [1,]: R:1, [1,]: R:1, [2,]: R:2, [2,]: R:2, [1,]: R:1, [1,]: R:1, [1,]: R:1, [DBI]: R:%60DBI%60 [0,]: R:0, [1,]: R:1, [1,]: R:1, [2,]: R:2, [2,]: R:2, [1,]: R:1, [1,]: R:1, [1,]: R:1,

References

https://martinfowler.com/eaaCatalog/repository.html

Super class

R6P::Singleton -> Repository

Methods

Public methods

AbstractRepository$new()

AbstractRepository$add()

AbstractRepository$del()

AbstractRepository$get()

Inherited methods

Method new()

Instantiate an object

Usage

AbstractRepository$new()

Method add()

Add an element to the Repository.

Usage

AbstractRepository$add(key, value)

Arguments

key

(character) Name of the element.

value

(?) Value of the element. Note: The values in the Repository are not necessarily of the same type. That depends on the implementation of AbstractRepository.

Method del()

Delete an element from the Repository.

Usage

AbstractRepository$del(key)

Arguments

key

(character) Name of the element.

Method get()

Retrieve an element from the Repository.

Usage

AbstractRepository$get(key)

Arguments

key

(character) Name of the element.